% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicLP.R
\name{maicLP}
\alias{maicLP}
\title{Checks if AD is within the convex hull of IPD using lp-solve}
\usage{
maicLP(ipd, ad)
}
\arguments{
\item{ipd}{a dataframe with n row and p column, where n is number of subjects and p is the number of variables used in matching.}

\item{ad}{a dataframe with 1 row and p column. The matching variables should be in the same order as that in \code{ipd}. The function does not check this.}
}
\value{
\item{lp.check }{0 = AD is inside IPD, and MAIC can be conducted; 2 = otherwise}
}
\description{
Checks if AD is within the convex hull of IPD using lp-solve
}
\examples{
## eIPD now contains response columns (r.cont, r.bin) in addition to the
## matching columns y1, y2. Subset to the matching columns explicitly.

## eAD[1,] is the scenario A in the reference paper,
## i.e. when AD is within IPD convex hull
maicLP(eIPD[, c('y1', 'y2')], eAD[1, c('y1', 'y2')])

## eAD[3,] is the scenario C in the reference paper,
## i.e. when AD is outside IPD convex hull
maicLP(eIPD[, c('y1', 'y2')], eAD[3, c('y1', 'y2')])
}
\references{
Glimm E and Yau L. (2022). 'Geometric approaches to assessing the numerical feasibility for conducting matching-adjusted indirect comparisons.' \emph{Pharmaceutical Statistics}, 21(5):974-987. \doi{10.1002/pst.2210}.
}
