% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osLP.2ipd.R
\name{osLP.2ipd}
\alias{osLP.2ipd}
\title{Checks whether two IPD datasets can be matched with lpSolve::lp}
\usage{
osLP.2ipd(ipd1, ipd2, mean.constraint = FALSE, catigorical.var = NULL)
}
\arguments{
\item{ipd1}{a dataframe with n1 row and p column, where n1 is number of subjects of the first IPD, and p is the number of variables used in standardization.}

\item{ipd2}{a dataframe with n2 row and p column, where n2 is number of subjects of the second IPD, and p is the number of variables used in standardization.}

\item{mean.constraint}{whether to restrict the weighted means to be within the ranges of observed means. Default is FALSE. When it is TRUE, there is a higher chance of not having a solution.}

\item{catigorical.var}{a list of variable names for the categorical variables in the data}
}
\value{
\item{lp.check}{0 = OS can be conducted; 2 = OS cannot be conducted}
}
\description{
Checks whether two IPD datasets can be matched with lpSolve::lp
}
\author{
Lillian Yau
}
