% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osLP.R
\name{osLP}
\alias{osLP}
\title{Checks whether two IPD datasets can be matched with lpSolve::lp}
\usage{
osLP(ipd1, ipd2)
}
\arguments{
\item{ipd1}{a dataframe with n1 row and p column, where n1 is number of subjects of the first IPD, and p is the number of variables used in standardization.}

\item{ipd2}{a dataframe with n2 row and p column, where n2 is number of subjects of the second IPD, and p is the number of variables used in standardization.}
}
\value{
\item{lp.check}{0 = OS can be conducted; 2 = OS cannot be conducted}
}
\description{
Checks whether two IPD datasets can be matched with lpSolve::lp
}
\author{
Lillian Yau
}
