% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osWt.2ipd.R
\name{osWt.2ipd}
\alias{osWt.2ipd}
\title{Optimal Standardization for two IPD's}
\usage{
osWt.2ipd(ipd1, ipd2, mean.constraint = FALSE, catigorical.var = NULL)
}
\arguments{
\item{ipd1}{a dataframe with n row and p column, where n is number of subjects and p is the number of variables used in matching.}

\item{ipd2}{the other IPD with the same number of columns}

\item{catigorical.var}{a list of variable names for the categorical variables in the data}
}
\value{
\item{osess.wt }{weights of optimal standardization by maximizing ESS. Scaled to .......}
\item{ipd.ess }{effective sample size. It is no smaller than the ESS given by the MAIC weights.}
\item{ipd.wtsumm}{weighted summary statistics of the matching variables after matching. they should be identical to the input AD when AD is within the IPD convex hull.}
}
\description{
Optimal Standardization for two IPD's
}
\details{
If dummy variables are already created for the categorical variables in the data set, and are present in \code{ipd1} and \code{ipd2}$, then \code{categorical.var} should be left as NULL.
}
