% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eAD.R
\docType{data}
\name{eAD}
\alias{eAD}
\title{three AD scenarios}
\format{
\describe{
\item{scen}{corresponds to scenarios A, B, and C in the reference manuscript (Glimm & Yau (2021)). Scenario A is very close to IPD center (see data(ipd)) and is within the IPD convex hull; scenario B is further away from IPD center but otherwise still inside the IPD convex hull; scenario C is outside IPD convex hull.}
\item{y1}{a numeric vector}
\item{y2}{a numeric vector}
}
}
\usage{
data(eAD)
}
\description{
Three artificial scenarios serves as the ad cases.
This is used in Glimm & Yau (2021)
}
\examples{
data(eAD)
}
\references{
Glimm & Yau (2021)
}
\keyword{datasets}
