\name{eIPD}
\alias{eIPD}
\docType{data}
\title{Example ipd set}
\description{Example ipd data}
\usage{data("eIPD")}
\format{
  A data frame with 42 observations on the following 2 variables. The illustrative example used in the reference manuscript (Glimm & Yau (2021)).
  \describe{
    \item{\code{y1}}{a numeric vector}
    \item{\code{y2}}{a numeric vector}
  }
}
\references{Glimm & Yau (2021). Geometric approaches to assessing the numerical feasibility for conducting matching-adjusted indirect comparisons. arXiv 2108.01896.
}
\examples{
head(eIPD)
}
\keyword{datasets}
