% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eIPD.R
\docType{data}
\name{eIPD}
\alias{eIPD}
\title{an IPD set}
\format{
\describe{
\item{\code{y1}}{a numeric vector}
\item{\code{y2}}{a numeric vector}
}
}
\usage{
data(eIPD)
}
\description{
an articial data set serves as the IPD set.
this is used in Glimm & Yau (2021)
}
\examples{
data(eIPD)
}
\references{
Glimm & Yau (2021)
}
\keyword{datasets}
