% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmWt.2ipd.R
\name{exmWt.2ipd}
\alias{exmWt.2ipd}
\title{Exact matching for two IPD's}
\usage{
exmWt.2ipd(ipd1, ipd2, mean.constrained = FALSE, catigorical.var = NULL)
}
\arguments{
\item{ipd1}{a dataframe with n row and p column, where n is number of subjects and p is the number of variables used in matching.}

\item{ipd2}{the other IPD with the same number of columns}

\item{mean.constrained}{whether to restrict the weighted means to be within the ranges of observed means. Default is FALSE. When it is TRUE, there is a higher chance of not having a solution.}

\item{catigorical.var}{a list of variable names for the categorical variables in the data}
}
\value{
\item{wt.1 }{weights of the exact matching by maximizing ESS for IPD 1}
\item{ess.1 }{effective sample size for IPD 1}
\item{wtd.means.1 }{weighted means of the matching variables for IPD 1. Same as wtd.means.2}
\item{wt.2 }{weights of the exact matching by maximizing ESS for IPD 2}
\item{ess.2 }{effective sample size for IPD 2}
\item{wtd.means.2 }{weghted means of the matching bariables for IPD 2. Same as wtd.means.1}
}
\description{
Exact matching for two IPD's
}
\details{
If dummy variables are already created for the categorical variables in the data set, and are present in \code{ipd1} and \code{ipd2}, then \code{categorical.var} should be left as NULL.
}
