% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicChecks-package.R
\docType{package}
\name{maicChecks-package}
\alias{maicChecks}
\alias{maicChecks-package}
\title{maicChecks: Exact Matching (ExM) and Matching-Adjusted Indirect Comparison (MAIC)}
\description{
The second version (0.2.0) contains implementation for exact matching which is an alternative to propensity score matching (see Glimm & Yau (2025)). The initial version (0.1.2) contains a collection of easy-to-implement tools for checking whether a MAIC can be conducted, as well as an alternative way of calculating weights (see Glimm & Yau (2021) \doi{10.1002/pst.2210}.)
}
\author{
\strong{Maintainer}: Lillian Yau \email{maicChecks@gmail.com}

Authors:
\itemize{
  \item Ekkehard Glimm
  \item Xinlei Deng \email{xinlei.deng.apha@gmail.com} (\href{https://orcid.org/0000-0001-8129-6007}{ORCID})
}

}
\keyword{internal}
