% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicChecks-package.R
\docType{package}
\name{maicChecks-package}
\alias{maicChecks}
\alias{maicChecks-package}
\title{maicChecks: Assessing the Numerical Feasibility for Conducting a Matching-Adjusted Indirect Comparison (MAIC)}
\description{
The initial version (0.1.2) contains a collection of easy-to-implement tools for checking whether a MAIC can be conducted. An alternative way of calculating weights is also included. These methods are introduced in Glimm & Yau (2021) \doi{10.1002/pst.2210}. The second version (0.2.0) adds an optimal standardisation as an alternative to propensity score matching method.
}
\author{
\strong{Maintainer}: Lillian Yau \email{maicChecks@gmail.com}

Authors:
\itemize{
  \item Ekkehard Glimm
}

}
\keyword{internal}
