% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim110.R
\docType{data}
\name{sim110}
\alias{sim110}
\title{Simulated data used for matching}
\format{
\describe{
\item{X1-X15}{Covariates used in matching.}
\item{Y}{Response variable}
}
}
\usage{
data(sim110)
}
\description{
sim110 is one of the simulated data presented in the simulation study in Glimm & Yau (2025).The covariates used in matching are X1 to X15. A response variable Y is simulated to depend on 6 of the 15 covariates.
}
\examples{
data(sim110)
}
\references{
Glimm & Yau (2021)
}
\keyword{datasets}
